Move ProgressBar in folder + split between budgeting and voting
