+++ front/app/containers/ProjectsShowPage/timeline/VotingResults/ProgressBar/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  xPicks: {
+    id: 'app.containers.ProjectsShowPage.timeline.VotingResults.ProgressBar.xPicks',
+    defaultMessage: '{picks, plural, one {1 pick} other {# picks}}',
+  },
+});
